modularize roux sdk
